refactor(gateway): extract workboard dal helpers
